+++ charmbuild/tests/__init__.py
@@ -1,0 +1,4 @@
+#!/usr/bin/env python3
+
+# Copyright 2025 Canonical Ltd.
+# See LICENSE file for licensing details.