+++ src/build.ts
@@ -14,6 +14,7 @@ import path from 'path'
 
 interface BuildCharmParams {
   plan: BuildPlan
+  buildContext: string
 }
 
 async function gitTreeId(p: string): Promise<string> {
@@ -28,11 +29,15 @@ async function gitTreeId(p: string): Promise<string> {
 }
 
 async function buildCharm(params: BuildCharmParams): Promise<void> {
-  core.startGroup('charmcraft pack')
-  await exec.exec('charmcraft', ['pack', '--verbosity', 'trace'], {
-    cwd: params.plan.source_directory,
-    env: { ...process.env, CHARMCRAFT_ENABLE_EXPERIMENTAL_EXTENSIONS: 'true' }
-  })
+  core.startGroup('charmbuild pack')
+  await exec.exec(
+    'charmbuild',
+    ['pack', '--verbosity', 'trace', '--build-context', params.buildContext],
+    {
+      cwd: params.plan.source_directory,
+      env: { ...process.env, CHARMCRAFT_ENABLE_EXPERIMENTAL_EXTENSIONS: 'true' }
+    }
+  )
   core.endGroup()
   const charmFiles = await (
     await glob.create(path.join(params.plan.source_directory, '*.charm'))
@@ -286,7 +291,7 @@ async function buildRock({
           .replace(/\.rock$/, '')
         const image = `ghcr.io/${github.context.repo.owner}/${plan.name}:${tree}-${base}`
         await exec.exec(
-          '/snap/rockcraft/current/bin/skopeo',
+          'rockcraft.skopeo',
           [
             '--insecure-policy',
             'copy',
@@ -326,7 +331,8 @@ export async function run(): Promise<void> {
     switch (plan.type) {
       case 'charm':
         await buildCharm({
-          plan
+          plan,
+          buildContext: core.getInput('build-context') || '.'
         })
         break
       case 'docker-image':